Update code consistency
replaced all instances of _Spacing consts to floats

+++ Assets/Scripts/Game/Projectiles/Enemy bullet systems/Leo/LeoBullet20.cs
@@ -4,7 +4,7 @@ using static MathHelper;
 public class LeoBullet20 : ScriptableEnemyBullet<LeoBulletSystem2, EnemyBullet>
 {
     const int BulletCount = 5;
-    const int BulletSpacing = 360 / BulletCount;
+    const float BulletSpacing = 360f / BulletCount;
 
     protected override float MaxLifetime => 2f;
